+++ CHANGELOG.md
@@ -5,6 +5,7 @@ under development
 -----------------
 - Add `Map`: an alias of map[string]interface{}.
 - Add `Decoder`, `Router.Decoder` and `Context.Decode` for decoding request's input.
+- Add `Context.Stringf`.
 
 v1.10.0 April 11, 2020
 ----------------------